+++ momo/src/main/java/ks52team01/admin/user/mapper/AdminMapper.java
@@ -1,22 +1,25 @@
-package ks52team01.admin.service;
+package ks52team01.admin.user.mapper;
 
-public interface AdminService {
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface AdminMapper {
 
 	public void adminPage();
-	
-	public void adminLoginMove() ;
-	
+
+	public void adminLoginMove();
+
 	public void adminLogoutAction();
-	
+
 	public void adminInfoModifyAction();
-	
+
 	public void adminJoinAction();
-	
+
 	public void adminInfoDetailAction();
-	
+
 	public void adminDeleteAction();
-	
+
 	public void adminFindIdAction();
-	
+
 	public void adminFindPwAction();
 }
