+++ momo/src/main/resources/mapper/admin/AdminMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
     PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
     "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="ks52team01.admin.mapper.AdminMapper">
+<mapper namespace="ks52team01.admin.user.mapper.AdminMapper">
